Swagger UI documentation is being returned under /?doc
Also code from swagger-ui.js is moved into root route file.
